refactor(jina): community 디렉토리로 이동 #52
