+++ src/entities/community/ui/DeleteModal/index.tsx
