+++ src/entities/reply/model/type.ts
@@ -1,4 +1,4 @@
-export interface Comment {
+export interface Reply {
   userId: string;
   rate: number; // 별점
   content: string;
