+++ src/features/community/createPost/lib/updateCheckedSet.ts
