+++ src/features/community/createPost/lib/useCreatePostMutation.ts
@@ -1,7 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
 import { useNavigate } from 'react-router-dom';
 
-import { fetchCreatePost } from 'entities/post/api/createPost';
+import { fetchCreatePost } from 'entities/community/api/createPost';
 
 import { ROUTES } from 'shared/config/routes';
