+++ src/features/community/createPost/lib/validate.ts
@@ -1,10 +1,10 @@
 import { ZodError } from 'zod';
 
-import { CreateFormSchema } from 'features/createPost/model/formSchema';
+import { CreateFormSchema } from 'features/community/createPost/model/formSchema';
 
+import { CreatePostDto } from 'entities/community/api/createPost';
 import { Genre } from 'entities/genre/model/genre';
 import { Mood } from 'entities/mood/model/mood';
-import { CreatePostDto } from 'entities/post/api/createPost';
 
 import { getYoutubeVideoId } from './youtubeId';
